retour au int

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: tboos <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/04 18:24:51 by tboos             #+#    #+#             */
-/*   Updated: 2016/03/16 14:34:58 by tboos            ###   ########.fr       */
+/*   Updated: 2016/03/16 14:40:03 by tboos            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -118,8 +118,8 @@ void				ft_striteri(char *s, void (*f)(unsigned int, char *));
 ** CHAR <-> INT
 */
 int					ft_atoi(const char *str);
-char				*ft_itoa(long long int n);
-char				*ft_st_itoa(long long int n);
+char				*ft_itoa(int n);
+char				*ft_st_itoa(int n);
 
 /*
 ** Padding et autre en attente printf
